Colocando as classes em arquivos proprios
