+++ Sources/Networking/DTOs/Events/EventDTO.swift
@@ -42,26 +42,6 @@ final public class EventDTO: Codable {
 }
 
 
-final public class LinkDTO: Codable {
-    public var id: UUID
-    public var text: String
-    
-    public init(id: UUID, text: String) {
-        self.id = id
-        self.text = text
-    }
-}
 
-final public class ChecklistItemDTO: Codable {
-    public var id: UUID
-    public var name: String
-    public var checked: Bool
-    public var targetAmount: Int
-    
-    public init(id: UUID, name: String, checked: Bool, targetAmount: Int) {
-        self.id = id
-        self.name = name
-        self.checked = checked
-        self.targetAmount = targetAmount
-    }
-}
+
+
